Ficha3

Description

This module contains Haskell definitions to calculate simple recursive function

Documentation

rembls

Arguments

:: [String]
List of words

-> Char
Charcter of reference

-> [String]
New list

This function removes all words from a list that starts with a given character

rembls  [] _ 
      = [] 
rembls (h:t) a 
      | head h == a = rembls t a   
      | otherwise 
       = h : rembls t 

Notes:

Expand

If the given list is empty, gives back an empty list

Exemples:

Expand

>>> rembls  ["João","Catarina"] 'C'
["João"] 
>>> rembls [] 'J'
[]

addFst

Arguments

:: [(Int, Int)]
List of pairs

-> Int
Number to add in the fisrt element of all pairs of the list

-> [(Int, Int)]
New List

This function adds a given nunber to all the first elements of the pairs in a list of pairs

addFst  [] _ 
      = []
addFst ((x,y):t) z 
      = (x + z, y) : addFst t z 

Notes:

Expand

If the list is empty gives back an empty list

Exemples:

Expand

>>> [(0,1),(1,2)] 1
[(1,1),(2,2)]
>>> [] 2
[]

bgg2

Arguments

:: [(Int, Int)]
List of pairs

-> Int
The biggest of all the sconde elements of the pairs of the list

This function calculates the biggest of all the second elements of a lsit of pairs

bgg2  []
    = error "Invalid List"
bgg2 [(x,y)]
    = y 
bgg2 ((x,y):(z,w):t)
    | y >= w = bgg2 ((x,y):t) 
    | otherwise 
     = bgg2 ((z,w):t) 

Notes:

Expand

If the list is empty gives back an error "Invalid List"

If the list has only one pair gives back the second element

Exemples:

Expand

>>> [(0,11),(1,20),(2,3)]
20
>>> [(0,11)]
11
>>> []
error "Invalid List"

type Name = String

type Coordinate = (Double, Double)

data Movement

Possible movementes

Constructors

N
North

S
South

E
East

W
West

type Movements = [Movement]

data PositionPersson

Indicates the possicion of someone

Constructors

Pos

Fields

Name

Name of the person

Coordinate

Coordinates of the person

positonM

Arguments

:: [PositionPersson]
List of people and their coordinates

-> Movement
Direction that they will move

-> [PositionPersson]
New list

This function makes the cordinates of evryone in a list move one unite to one direction

positonM [] _ 
        = [] 
If they move to North (Up)            
positonM ((Pos n (x,y)):t) N 
        = Pos n (x,y + 1) : positonM t N  
If they move to South (Down)
positonM ((Pos n (x,y)):t) S 
        = Pos n (x,y - 1) : positonM t S 
If they move to East (Right) 
positonM ((Pos n (x,y)):t) E 
        = Pos n (x + 1,y) : positonM t E 
If they move to West (Left)
positonM ((Pos n (x,y)):t) W 
        = Pos n (x - 1,y) : positonM t W  

Notes:

Expand

If the list is empty gives back an empty list

Exemples:

Expand

>>> [Pos "Joao" (1,1),Pos "Catarina" (2,2)] N
[Pos "Joao" (1.0,2.0),Pos "Catarina" (2.0,3.0)]
>>> [Pos "Joao" (1,1),Pos "Catarina" (2,2)] S
[Pos "Joao" (1.0,0.0),Pos "Catarina" (2.0,1.0)]
>>> [Pos "Joao" (1,1),Pos "Catarina" (2,2)] E
[Pos "Joao" (2.0,1.0),Pos "Catarina" (3.0,2.0)]
>>> [Pos "Joao" (1,1),Pos "Catarina" (2,2)] W
[Pos "Joao" (0.0,1.0),Pos "Catarina" (1.0,2.0)]
>>> [] N
[]

multPstn

Arguments

:: PositionPersson
The person and their coordinates

-> Movements
List of movementes

-> PositionPersson
The person with the new coodinates

This function move a person a certain amount of times in any direction

multPstn x [] 
        = x      
multPstn (Pos n (x,y)) (h:t) 
       = multPstn (mov (Pos n (x,y)) h) t    
Auxiliary function that calculates the movement in the head of the list 
  where 
       If it moves North (Up)
        mov (Pos n (x,y)) N 
           = Pos n (x,y + 1) 
       If it moves South (Down)
        mov (Pos n (x,y)) S 
           = Pos n (x,y - 1) 
       If it moves East (Right)
        mov (Pos n (x,y)) E 
           = Pos n (x + 1,y) 
       If it moves West (Left)
        mov (Pos n (x,y)) W 
           = Pos n (x - 1,y)  

Notes:

Expand

If the list of movements is empty gives back the same coordinates

Exemples:

Expand

>>> (Pos "Joao" (0,0)) [N,E,E]
Pos "Joao" (2.0,1.0)
>>> (Pos "Joao" (0,0)) [S,S,W]
Pos "Joao" (-1.0,-2.0) 
>>> (Pos "Joao" (0,0)) []
Pos "Joao" (0,0) 

multPstnM

Arguments

:: [PositionPersson]
List of people and their coordinates

-> Movements
List of movements

-> [PositionPersson]
List of people and their coordinates after moving

This function moves a list of people a certain amount of times in any direction

multPstnM  [] _ 
         = [] 
multPstnM  l []  
         = l 
multPstnM (Pos n (x,y):t) (h:t1) 
         = multPstn (Pos n (x,y)) (h:t1) : multPstnM t (h:t1)

Notes:

Expand

If the list of people is empty gives back an empty list

If the list of movements is empty gives back the same list of people

Exemples:

Expand

>>> [Pos "Joao" (0,0), Pos "Catarina" (0,0)] [S,S,W]
[Pos "Joao" (-1.0,-2.0),Pos "Catarina" (-1.0,-2.0)]
>>> [] [S,S,W]
[]
>>> [Pos "Joao" (0,0), Pos "Catarina" (0,0)] []
[Pos "Joao" (0,0), Pos "Catarina" (0,0)] 

personN

Arguments

:: [PositionPersson]
List of people and their coordinates

-> Maybe PositionPersson
The person more to the north (Bigger y coordinate)

This function indentifies the person positioned more to the north

personN [] 
       = Nothing 
personN [x] 
       = Just x 
personN (Pos n1 (x1,y1): Pos n2 (x2,y2): t) 
       = Just (north (Pos n1 (x1,y1):Pos n2 (x2,y2):t))    
  where 
Auxiliar function that indentifies the person position more to the north (Bigger y coordinate)/  
        north [x] 
             = x  
        north (Pos n1 (x1,y1):Pos n2 (x2,y2):t)
             | y1 >= y2 = north (Pos n1 (x1,y1):t) 
             | otherwise 
              = north (Pos n2 (x2,y2):t)

Notes:

Expand

If the list is empty gives back nothing

If the list has only one element, gives back that same element

Exemples:

Expand

>>> [Pos "Joao" (0,1), Pos "Catarina" (0,2)]
Just (Pos "Catarina" (0.0,2.0))
>>> [Pos "Joao" (0,1)]
Just (Pos "Joao" (0.0,1.0))
>>> personN []
Nothing 

peopleN

Arguments

:: [PositionPersson]
List of people and their coordinates

-> [Name]
List of names of everyone more to the north

This functions gives the list of names of everyone that is more to the north

peopleN  []
       = []  
peopleN (Pos n (x,y):t) 
       = listN (Pos n (x,y):t) maxY 
  where 
Auxiliar function that tranformas the type of the result the function maxN into a Double
        maxY = maxN (Pos n (x,y):t) 
        --
Auxiliar function that calculates the coordinates of the person more to the north (Bigger y coordinate)        
        maxN :: 
            [PositionPersson] (List of people and their coordinates) 
         -> Double            (Coordinates of the people more to the south)
        maxN [Pos n (x,y)]
            = y
        maxN (Pos n1 (x1,y1):Pos n2 (x2,y2):t)         
            | y1 >= y2 = maxN (Pos n1 (x1,y1):t) 
            | otherwise 
             = maxN (Pos n2 (x2,y2):t) 
        --
Auxiliar function that makes the list of names of everyone in a certain coordinate  
        listN :: 
            [PositionPersson] (List of people and their coordinates) 
         -> Double            (Coordinates of the people more to the south)
         -> [Name]            (List of names of everyone more to the north) 
        listN  [] c 
             = []   
        listN (Pos n (x,y):t) c 
             | y == c = n : listN t c 
             | otherwise 
              = listN t c

Notes:

Expand

If the list of people is empty than it gives back an empty list

If thres only one element than it gives back an list with that persons name

Exemples:

Expand

>>> [Pos "Joao" (1.0, 5.0), Pos "Catarina" (2.0, 10.0),Pos "Carvalhosa" (4.0, 10.0)]
["Catarina","Carvalhosa"]
>>> [Pos "Joao" (1.0, 5.0)]
["Joao"]
>>> []
[] 

ntoRight

Arguments

:: [a]
List

-> Int
How many times the elements move to the right

-> [a]
List after the elements moved

This function moves all elementes of a list a certain number of times to the right

ntoRight  l 0 
        = l 
ntoRight  [] _
        = [] 
ntoRight (h:t) n 
        = ntoRight (last t : h : init t) (n - 1)

Notes:

Expand

If the list is empty gives back an empty list

If the number gib«ven is 0 gives back the same list

The elements in the end of the list move to the begning

Exemples:

Expand

>>> [1,2,3,4,5] 2
[4,5,1,2,3]
>>> [1,2,3,4,5] 0
[1,2,3,4,5]
>>> [] 1
[]